Shop UI Implemented and tested
The shop UI is designed and implemented and tested (as far as i can go atm) .
Purchasing is not yet implemented as we do not have a currency system yet but i have started on a way to implement it by using playerprefs and events. This simply requires currency and Upgrade unlocking to complete.
Shop NPC is implemented by not yet tested as we do not have interaction yet